Indulge In Your Learning
I made this is using  https://giphy.com and https://spacetypegenerator.com

This is a quote from Iris Bildstein.  She teaches Curriculum Design at Teachers College, Columbia University.  I often think about this quote when I am feeling stressed about the amount of work that I have to do.  I want to remember to “indulge in my learning” and be aware of the fact that it is a privilege to be a student.